requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.5
djangorestframework==3.14.0
Pillow==9.4.0
pycodestyle==2.10.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
